+++ src/Exceptions/ListenException.cpp
@@ -1,6 +1,6 @@
 #include "ListenException.hpp"
 
-ListenException::ListenException(u_int16_t socket_fd) : ServerException(-1, socket_fd) {}
+ListenException::ListenException(u_int16_t socket_fd) : ServerException((u_int16_t)-1, socket_fd) {}
 
 ListenException::~ListenException() {}
